+++ sdsp2l_algo/interface.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Export funciton interface
-extern "C" __declspec(dllexport) long lp2lHitCnt();
+extern "C" __declspec(dllexport) int lp2lHitCnt();
 extern "C" __declspec(dllexport) long lpl2ChkCnt();
 extern "C" __declspec(dllexport) int iGetDevCap();
 extern "C" __declspec(dllexport) int iGetDdrSize();
@@ -13,5 +13,5 @@ extern "C" __declspec(dllexport) int iGetTableSize();
 
 extern "C" __declspec(dllexport) int* iGetEraseCntTable(int ch);
 extern "C" __declspec(dllexport) int* iGetReadCntTable(int ch);
-extern "C" __declspec(dllexport) int iGetIoBurstCnt(int ch);
+extern "C" __declspec(dllexport) long iGetIoBurstCnt(int ch);
 extern "C" __declspec(dllexport) int iClearChkHitCnt();
